% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicfuns.R
\name{change_log}
\alias{change_log}
\title{Calculate change logic}
\usage{
change_log(x, y, x1, y1, x2, y2)
}
\arguments{
\item{x}{numeric, x coordinate of data point}

\item{y}{numeric, y coordinate of data point}

\item{x1}{numeric, x coordinate of min point of zone boundary line}

\item{y1}{numeric, y coordinate of min point of zone boundary line}

\item{x2}{numeric, x coordinate of max point of zone boundary line}

\item{y2}{numeric, y coordinate of max point of zone boundary line}
}
\value{
logic
}
\description{
logic whether to change a data points zone label. Points on line are default given to inner zone
}
\examples{
## Not Run
#change_log(data$x, data$y, line$x[ii], line$y[ii], line$x2[ii], line$y2[ii])
## End Not Run
}
