% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicfuns.R
\name{change_log_out}
\alias{change_log_out}
\title{Calculate change logic outer direction}
\usage{
change_log_out(x, y, x1, y1, x2, y2, minx)
}
\arguments{
\item{x}{numeric, x coordinate of data point}

\item{y}{numeric, y coordinate of data point}

\item{x1}{numeric, x coordinate of min point of zone boundary line}

\item{y1}{numeric, y coordinate of min point of zone boundary line}

\item{x2}{numeric, x coordinate of max point of zone boundary line}

\item{y2}{numeric, y coordinate of max point of zone boundary line}

\item{minx}{numeric, min x value of grid}
}
\value{
A logic value
}
\description{
Calculate logic whether to change a data points zone label. Points on line are given to outter zone
}
\examples{
## Not Run
#change_log_out(data$x, data$y, line$x[ii], line$y[ii], line$x2[ii], line$y2[ii], minx)
## End Not Run
}
