% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_line_cross}
\alias{check_line_cross}
\title{Check Line Cross}
\usage{
check_line_cross(coords)
}
\arguments{
\item{coords}{A dataframe, coordinate points of lines for Custom Error Grid}
}
\value{
Error message if lines cross. Otherwise nothing.
}
\description{
Check if lines of Custom Error Grid cross. Lines can overlap, but they cannot intersect.
}
\examples{
\dontshow{
coordinates <-  data.frame(x = c(0,1,3,4,4,
0,2,2,5,5,7,
2,2,4,6,6,8,8,10,
6,8,8,10),
y = c(7,7,9,9,10,
     3,5,7,7,8,10,
     0,1,1,3,4,4,6,8,
     0,2,3,3),
line_id=c(1,1,1,1,1,
         2,2,2,2,2,2,
         3,3,3,3,3,3,3,3,
         4,4,4,4))
}
check_line_cross(coordinates)
}
