% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_lines}
\alias{check_lines}
\title{Check Lines}
\usage{
check_lines(coords)
}
\arguments{
\item{coords}{Coordinate points of lines for Custom Error Grid}
}
\value{
Error message if lines are not entered correctly. Otherwise nothing.
}
\description{
Check if lines have been entered in correct format
}
\examples{
\dontshow{
coordinates <-  data.frame(x = c(0,1,3,4,4,
0,2,2,5,5,7,
2,2,4,6,6,8,8,10,
6,8,8,10),
y = c(7,7,9,9,10,
     3,5,7,7,8,10,
     0,1,1,3,4,4,6,8,
     0,2,3,3),
line_id=c(1,1,1,1,1,
         2,2,2,2,2,2,
         3,3,3,3,3,3,3,3,
         4,4,4,4))
}
check_lines(coordinates)
}
