% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Clarkes_coord.R
\docType{data}
\name{Clarkes_coord}
\alias{Clarkes_coord}
\title{Clarke's Error Grid coordinates (blood glucose)}
\format{
An RData object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7196876/}
}
\usage{
data(Clarkes_coord)
}
\description{
The coordinates data file to use Clarke's Error Grid for blood glucose
}
\examples{
data(Clarkes_coord)
print(Clarkes_coord)
}
\keyword{datasets}
