% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Clarkes_arealabs.R
\docType{data}
\name{Clarkes_labs}
\alias{Clarkes_labs}
\title{Clarke's Error Grid for blood glucose area labels}
\format{
An RData object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7196876/}
}
\usage{
data(Clarkes_labs)
}
\description{
The area labels data file to use Clarke's Error Grid for blood glucose
}
\examples{
data(Clarkes_labs)
print(Clarkes_labs)
}
\keyword{datasets}
