% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_datapoints.R
\docType{data}
\name{datapoints}
\alias{datapoints}
\title{Data for examples}
\format{
An RData object
}
\usage{
data(datapoints)
}
\description{
A basic example set of data points to use with the example Error Grid
}
\examples{
data(datapoints)
print(datapoints)
}
\keyword{datasets}
