% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Eleydata.R
\docType{data}
\name{Eley}
\alias{Eley}
\title{Eley et al. Blood Pressure Data}
\format{
An RData object
}
\usage{
data(Eley)
}
\description{
Blood pressure data from three different methods: Invasive arterial line (gold standard), non-Invasive cuff, Clearsite(tm).
The methods used in this package were developed for the analysis of this data.
The data contains measurments from a number of patients at numerous time points (0 up to 60 minutes).
Both systolic arterial pressure and mean arterial pressure are included.
The Error Grid from Saugel et al. (2018) Anesth Analg was used.
}
\examples{
data(Eley)
print(Eley)
}
\keyword{datasets}
