% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getzones.R
\name{get_zones}
\alias{get_zones}
\title{Calculate the zones in the Custom Error Grid}
\usage{
get_zones(data, coords, area_labels, dir = "in")
}
\arguments{
\item{data}{A dataframe, dataset of points of (gold standard method, novel method)}

\item{coords}{A dataframe, coordinate points of lines for Custom Error Grid}

\item{area_labels}{A vector of area labels for each zone}

\item{dir}{A character value for direction. Should points that sit on a Error Grid line go to the inner or outer zone c("in","out")}
}
\value{
A vector of zones. You can add this vector to a dataframe, or keep it as a separate object.
        It does not automatically add it to the dataset the data points came from.
}
\description{
Calculate the zones in the Custom Error Grid for the provided data points.
This function does not perform checks on the entered data and Error Grid coordinates
}
\examples{
\dontshow{
datapoints <- data.frame(x = c(1, 2, 4, 6, 5, 5, 5, 3, 4, 9, 8, 8, 5, 3, 2, 2, 7, 8, 9, 9, 1, 2, 2, 1, 4, 6, 9, 3, 4,10, 0, 5, 9, 6,10,10, 0, 7, 0, 8, 5, 2, 4, 8),
y = c(9, 8, 8, 9, 7, 8, 5, 3, 6, 7, 4, 5, 2, 1, 2, 0, 1, 3, 1, 3,10, 7, 5, 4, 9, 4, 5, 7,10, 8, 5, 0,10, 0, 3, 6, 7,10, 0, 2,10, 6, 1, 6))
coordinates <-  data.frame(x = c(0,1,3,4,4,
0,2,2,5,5,7,
2,2,4,6,6,8,8,10,
6,8,8,10),
y = c(7,7,9,9,10,
     3,5,7,7,8,10,
     0,1,1,3,4,4,6,8,
     0,2,3,3),
line_id=c(1,1,1,1,1,
         2,2,2,2,2,2,
         3,3,3,3,3,3,3,3,
         4,4,4,4))
}
get_zones(data = datapoints,
          coords = coordinates,
          area_labels = c("C1","B1","A","B2","C2"))
}
