% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MAP_coord.R
\docType{data}
\name{MAP_coord}
\alias{MAP_coord}
\title{Coordinates for Saugel MAP Error Grid for blood glucose}
\format{
An RData object
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29239945/}
}
\usage{
data(MAP_coord)
}
\description{
The coordinates data file to Saugel MAP Error Grid for blood glucose
}
\examples{
data(MAP_coord)
print(MAP_coord)
}
\keyword{datasets}
