% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MAP_arealabs.R
\docType{data}
\name{MAP_labs}
\alias{MAP_labs}
\title{Area labels for Saugel MAP Error Grid for blood glucose}
\format{
An RData object
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29239945/}
}
\usage{
data(MAP_labs)
}
\description{
The area labels data file to use Saugel MAP Error Grid for blood glucose
}
\examples{
data(MAP_labs)
print(MAP_labs)
}
\keyword{datasets}
