% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ParkesT1_coord.R
\docType{data}
\name{Parkes_T1_coord}
\alias{Parkes_T1_coord}
\title{Parkes Error Grid coordinates  for blood glucose Type 1 diabetes}
\format{
An RData object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3876371/}
}
\usage{
data(Parkes_T1_coord)
}
\description{
The coordinates data file to use Parkes Error Grid for blood glucose in Type 1 diabetes
}
\examples{
data(Parkes_T1_coord)
print(Parkes_T1_coord)
}
\keyword{datasets}
