% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ega.R
\name{plot_EGA}
\alias{plot_EGA}
\title{Plot data in an Error Grid Analysis}
\usage{
plot_EGA(
  data = NULL,
  coords,
  x_lab = "x",
  y_lab = "y",
  x = NULL,
  y = NULL,
  zones = NULL,
  area_labels = NULL,
  check_zone = FALSE
)
}
\arguments{
\item{data}{dataframe, data to plot in zones}

\item{coords}{dataframe, coordinates of lines to make up the Error Grid Analysis (required)}

\item{x_lab}{character, label for x axis}

\item{y_lab}{character, label for y axis}

\item{x}{vector, x coordinate of data points to plot}

\item{y}{vector, x coordinate of data points to plot}

\item{zones}{vector, zones that each data point is in}

\item{area_labels}{vector, label for each zone}

\item{check_zone}{TRUE or FALSE, whether to check the zones have been calculated correctly}
}
\value{
A ggplot object. Will either be displayed in plot window, or can be assigned to an object.
        More ggplot layers can be added to the object afterwards.
}
\description{
Performs zone assignment and plots data coloured by zone.
}
\examples{
\dontshow{
 datapoints <- data.frame(x = c(1, 2, 4, 6, 5, 5, 5, 3, 4, 9, 8, 8, 5, 3, 2, 2, 7, 8, 9, 9, 1, 2, 2, 1, 4, 6, 9, 3, 4,10, 0, 5, 9, 6,10,10, 0, 7, 0, 8, 5, 2, 4, 8),
y = c(9, 8, 8, 9, 7, 8, 5, 3, 6, 7, 4, 5, 2, 1, 2, 0, 1, 3, 1, 3,10, 7, 5, 4, 9, 4, 5, 7,10, 8, 5, 0,10, 0, 3, 6, 7,10, 0, 2,10, 6, 1, 6))
coordinates <-  data.frame(x = c(0,1,3,4,4,
0,2,2,5,5,7,
2,2,4,6,6,8,8,10,
6,8,8,10),
y = c(7,7,9,9,10,
     3,5,7,7,8,10,
     0,1,1,3,4,4,6,8,
     0,2,3,3),
line_id=c(1,1,1,1,1,
         2,2,2,2,2,2,
         3,3,3,3,3,3,3,3,
         4,4,4,4))
}
# zones not yet calculated
plot_EGA(data = datapoints, coords = coordinates,
         x_lab = "my gold standard", y_lab = "my new measure",
         area_labels = c("C1","B1","A","B2","C2"))
# zones calculated (data is then 3 columns - x,y,zone)
datapoints$zone <- zones(data = datapoints, coords = coordinates,
                         area_labels = c("C1","B1","A","B2","C2"))
plot_EGA(data = datapoints, coords = coordinates,
         x_lab = "my gold standard", y_lab = "my new measure")
}
