% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SAP_coord.R
\docType{data}
\name{SAP_coord}
\alias{SAP_coord}
\title{Coordinates for Saugel SAP Error Grid for blood glucose}
\format{
An RData object
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29239945/}
}
\usage{
data(SAP_coord)
}
\description{
The coordinates data file to Saugel SAP Error Grid for blood glucose
}
\examples{
data(SAP_coord)
print(SAP_coord)
}
\keyword{datasets}
