% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperleftarea.R
\name{upplftarea}
\alias{upplftarea}
\title{Upper Left Area}
\usage{
upplftarea(df, ymin, ymax, xmin, xmax)
}
\arguments{
\item{df}{A dataframe with variables x and y,}

\item{ymin}{numeric, minimum y of grid}

\item{ymax}{numeric, maximum y of grid}

\item{xmin}{numeric, minimum x of grid}

\item{xmax}{numeric, maximum x of grid}
}
\value{
Numeric area above zone boundary in the grid
}
\description{
Calculates the upper left area of each zone boundary. That is, the area made by a Error Grid line and the top left corner of the Grid.
}
\examples{
\dontrun{
upplftarea(df,ymin,ymax,xmin,xmax)
}
}
