% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicfuns.R
\name{change_log_ver}
\alias{change_log_ver}
\title{Calculate change logic on vertical zone boundaries. Favouring inner direciton if points on line.}
\usage{
change_log_ver(x, y, x12, y1, y2)
}
\arguments{
\item{x}{numeric, x coordinate of data point}

\item{y}{numeric, y coordinate of data point}

\item{x12}{numeric, x coordinate of zone boundary line}

\item{y1}{numeric, y coordinate of min point of zone boundary line}

\item{y2}{numeric, y coordinate of max point of zone boundary line}
}
\value{
A logic value
}
\description{
Calculate change logic on vertical zone boundaries. Favouring inner direciton if points on line.
}
\examples{
## Not Run
#change_log_ver(data$x, data$y, line$x[ii], line$y[ii], line$y2[ii])
## End Not Run
}
