% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_coords.R
\docType{data}
\name{coordinates}
\alias{coordinates}
\title{Data for examples}
\format{
An RData object
}
\usage{
data(coordinates)
}
\description{
A basic example set of cordinates for an Error Grid
}
\examples{
data(coordinates)
print(coordinates)
}
\keyword{datasets}
