% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_MAP_coord.R
\docType{data}
\name{MAP_coord}
\alias{MAP_coord}
\title{Saugel MAP Error Grid Coordinates}
\format{
An RData object
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29239945/}
}
\usage{
data(MAP_coord)
}
\description{
The coordinates data file to the Saugel et al (2018) mean arterial blood pressure (MAP) Error Grid
}
\examples{
data(MAP_coord)
print(MAP_coord)
}
\keyword{datasets}
