% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ParkesT1_arealabs.R
\docType{data}
\name{Parkes_T1_labs}
\alias{Parkes_T1_labs}
\title{Parkes Error Grid for blood glucose in Type 1 diabetes area labels}
\format{
An RData object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3876371/}
}
\usage{
data(Parkes_T1_labs)
}
\description{
The area labels data file to use Parkes Error Grid for blood glucose in Type 1 diabetes
}
\examples{
data(Parkes_T1_labs)
print(Parkes_T1_labs)
}
\keyword{datasets}
