% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ParkesT2_arealabs.R
\docType{data}
\name{Parkes_T2_labs}
\alias{Parkes_T2_labs}
\title{Area labels for Parkes Type 2 Error Grid for blood glucose}
\format{
An RData object
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3876371/}
}
\usage{
data(Parkes_T2_labs)
}
\description{
The area labels data file to use Parkes Type 2 Error Grid for blood glucose
}
\examples{
data(Parkes_T2_labs)
print(Parkes_T2_labs)
}
\keyword{datasets}
