% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_SAP_arealabs.R
\docType{data}
\name{SAP_labs}
\alias{SAP_labs}
\title{Area labels for Saugel SAP Error Grid for blood glucose}
\format{
An RData object
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/29239945/}
}
\usage{
data(SAP_labs)
}
\description{
The area labels data file to use Saugel SAP Error Grid for blood glucose
}
\examples{
data(SAP_labs)
print(SAP_labs)
}
\keyword{datasets}
