% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upperleftarea.R
\name{upplftarea}
\alias{upplftarea}
\title{Upper Left Area}
\usage{
upplftarea(df, ymin, ymax, xmin, xmax)
}
\arguments{
\item{df}{dataframe with variables x and y,}

\item{ymin}{numeric, minimum y of grid}

\item{ymax}{numeric, maximum y of grid}

\item{xmin}{numeric, minimum x of grid}

\item{xmax}{numeric, maximum x of grid}
}
\value{
numeric area above zone boundary
}
\description{
Upper Left Area of each zone boundary
}
\examples{
\dontrun{
upplftarea(df,ymin,ymax,xmin,xmax)
}
}
